requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
Django==1.11.15
djangorestframework==3.8.2
fuzzywuzzy==0.17.0
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
nltk==3.3
numpy==1.15.2
pandas==0.23.4
pkg-resources==0.0.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-Levenshtein==0.12.0
pytz==2018.5
requests==2.19.1
six==1.11.0
subprocess32==3.5.2
urllib3==1.23
